requirements.txt
# Core ML dependencies
torch>=2.0.0
open-clip-torch>=2.24.0
faiss-cpu>=1.7.4

# Image processing
Pillow>=10.0.0
imagehash>=4.3.0

# CLI
click>=8.0.0
rich>=13.0.0

# Web API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Utilities
numpy>=1.24.0

